Finance/TradingBot/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
gevent==20.4.0
greenlet==0.4.15
idna==2.9
python-dateutil==2.8.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9
websocket==0.2.1
websocket-client==0.57.0
